requirements.txt
Django==4.0.6
django_filter==22.1
djangorestframework==3.13.1
